Uniform usage of NavBar page object + specific navbar for admin/typist
