+++ frontend/e2e-tests/page-objects/data_entry/DataEntryBasePgObj.ts
@@ -1,11 +1,9 @@
 import { type Locator, type Page } from "@playwright/test";
 
-import { NavBar } from "../NavBarPgObj";
 import { ProgressList } from "./ProgressListPgObj";
 import { UnsavedChangesModal } from "./UnsavedChangesModalPgObj";
 
 export class DataEntryBasePage {
-  readonly navBar: NavBar;
   readonly unsavedChangesModal: UnsavedChangesModal;
   readonly progressList: ProgressList;
 
@@ -16,7 +14,6 @@ export class DataEntryBasePage {
   readonly feedbackHeader: Locator;
 
   constructor(protected readonly page: Page) {
-    this.navBar = new NavBar(page);
     this.unsavedChangesModal = new UnsavedChangesModal(page);
     this.progressList = new ProgressList(page);
