+++ frontend/e2e-tests/page-objects/election/OverviewPgObj.ts
@@ -1,9 +1,6 @@
 import { type Locator, type Page } from "@playwright/test";
 
-import { NavBar } from "../NavBarPgObj";
-
 export class OverviewPgObj {
-  readonly navBar: NavBar;
   readonly main: Locator;
   readonly header: Locator;
   readonly alert: Locator;
@@ -11,7 +8,6 @@ export class OverviewPgObj {
   readonly elections: Locator;
 
   constructor(protected readonly page: Page) {
-    this.navBar = new NavBar(page);
     this.main = page.getByRole("main");
     this.header = page.getByRole("heading", { name: "Beheer verkiezingen" });
     this.alert = page.getByRole("heading", { name: "Je account is ingesteld" });
