+++ frontend/e2e-tests/tests/data-entry/data-entry.valid.model.e2e.ts
@@ -5,6 +5,7 @@ import { DataEntryHomePage } from "e2e-tests/page-objects/data_entry/DataEntryHo
 import { DifferencesPage } from "e2e-tests/page-objects/data_entry/DifferencesPgObj";
 import { ExtraInvestigationPage } from "e2e-tests/page-objects/data_entry/ExtraInvestigationPgObj";
 import { VotersAndVotesPage } from "e2e-tests/page-objects/data_entry/VotersAndVotesPgObj";
+import { TypistNavBar } from "e2e-tests/page-objects/nav_bar/TypistNavBarPgObj";
 import { createMachine } from "xstate";
 
 import { ExtraInvestigation, VotersCounts, VotesCounts } from "@/types/generated/openapi";
@@ -187,6 +188,7 @@ test.describe("Data entry model test - valid data", () => {
         const votersAndVotesPage = new VotersAndVotesPage(page);
         const differencesPage = new DifferencesPage(page);
         const abortModal = new AbortInputModal(page);
+        const navBar = new TypistNavBar(page);
 
         const pollingStationsPageStates = {
           pollingStationsPageDiscarded: async () => {
@@ -309,7 +311,7 @@ test.describe("Data entry model test - valid data", () => {
             await votersAndVotesPage.abortInput.click();
           },
           NAV_TO_POLLING_STATION_PAGE: async () => {
-            await votersAndVotesPage.navBar.clickElection(election.election.location, election.election.name);
+            await navBar.clickElection(election.election.location, election.election.name);
           },
           SUBMIT: async () => {
             await votersAndVotesPage.next.click();
